Updated CRUD operations, added search functionality, and upgraded VS version
In this commit, several significant changes have been made. The `CRUD application 2.csproj` file was updated to include and then remove a reference to `Microsoft.CodeDom.Providers.DotNetCompilerPlatform`. A no-op change was also made to the closing `Project` tag. The `CRUD application 2.sln` file was updated to upgrade the Visual Studio version from 16 to 17 and a new project `UserControllerTest` was added. The `UserController.cs` file was updated to implement CRUD operations on the `User` model and a new `Search` method was added. Lastly, the `Index.cshtml` file was updated to add a form for the new `Search` functionality and to change the format of the `ActionLink` calls in the user list table.